fixed(TMChessServer) : fixed submitMove while returning instead of false return MoveResponse type object

+++ TMChessServer/src/main/java/com/thinking/machines/chess/server/TMChessServer.java
@@ -252,7 +252,7 @@ byte playerColor=move.player;
 //update the move in list
 game.moves.add(move);
 //switch the player
-game.activePlayer=(playerColor==1)?0:1;
+game.activePlayer=(byte)((playerColor==1)?0:1);
 return moveResponse;
 }
